+++ src/main/java/templates/CourseView.java
@@ -315,7 +315,7 @@ public class CourseView extends View {
                 totalScore += score;
             }
             //undergraduatePanel.add(new JLabel(Double.toString(totalScore)));
-            undergraduatePanel.add(new JLabel(Float.toString(student.getGrade(course.getId()))));
+            undergraduatePanel.add(new JLabel(Double.toString(student.getGrade(course.getId()))));
         }
 
         // adding in the graduate stuff
